Refactor Screenshot files and handle errors
