+++ Screenshots/Classes/ScreenshotError.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public enum ScreenshotError: Error {
+  case userCancelled
+  case terminationStatusNotZero(_ status: Int, _ outputData: Data)
+//  case invalidImage
+//  case missingMetadataRectProperty
+}
